MAR-144: Move history classes into separate module
