+++ api/barriers/history/__init__.py
@@ -1,0 +1,4 @@
+from .assessments import AssessmentHistoryFactory
+from .barriers import BarrierHistoryFactory
+from .notes import NotesHistoryFactory
+from .team_members import TeamMemberHistoryFactory
