+++ api/barriers/history/notes.py
@@ -1,0 +1,12 @@
+from .base import BaseHistoryItem, HistoryItemFactory
+
+
+class NoteTextHistoryItem(BaseHistoryItem):
+    field = "text"
+
+
+class NotesHistoryFactory(HistoryItemFactory):
+    class_lookup = {}
+    history_item_classes = (
+        NoteTextHistoryItem,
+    )
