+++ api/core/utils.py
@@ -23,12 +23,6 @@ def cleansed_username(user):
     return None
 
 
-def get_changed_fields(new_record, old_record):
-    if hasattr(new_record, "get_changed_fields"):
-        return new_record.get_changed_fields(old_record)
-    return new_record.diff_against(old_record).changed_fields
-
-
 class EchoUTF8:
     """
     Writer that echoes written data and encodes to utf-8 if necessary.
